refactor(events-page): recategorise related files
PT-1213
