+++ src/domain/events/eventsPage.module.scss
@@ -22,11 +22,6 @@
     font-size: var(--fontsize-heading-m);
   }
 
-  .eventCount {
-    font-weight: normal;
-    margin-left: 0.5rem;
-  }
-
   .comingEventsTitleWrapper {
     display: grid;
 
@@ -48,21 +43,5 @@
     @include respond_above(sm) {
       grid-template-columns: var(--width-input) 1fr auto;
     }
-  }
-
-  .eventsContainer {
-    display: grid;
-    grid-gap: var(--grid-gap);
-    margin: var(--margin-events-container);
-  }
-
-  .showMoreButton {
-    text-align: center;
-    color: var(--color-show-more-button);
-    font-size: var(--fontsize-show-more-button);
-    font-weight: var(--fontweight-show-more-button);
-    text-decoration: none;
-    margin: 0 auto;
-    display: block;
   }
 }
